+++ internal/repository/postgresql/migrations/000009_create_feed_preferences_table.down.sql
@@ -1,0 +1,5 @@
+-- Copyright (c) VirtualTam
+-- SPDX-License-Identifier: MIT
+
+DROP TABLE IF EXISTS feed_preferences;
+DROP TYPE IF EXISTS feed_entry_visibility;