+++ internal/repository/postgresql/pgfeed/models.go
@@ -183,6 +183,12 @@ func (f *DBSubscribedFeed) asSubscribedFeed() feedquerying.SubscribedFeed {
 	}
 }
 
+type DBPreferences struct {
+	UserUUID    string    `db:"user_uuid"`
+	ShowEntries string    `db:"show_entries"`
+	UpdatedAt   time.Time `db:"updated_at"`
+}
+
 type DBSubscription struct {
 	UUID         string `db:"uuid"`
 	CategoryUUID string `db:"category_uuid"`